python/requirements.txt
scalekit-sdk-python
anthropic
requests
python-dotenv
